Sort of working bfcl
I decided to not reroute everything into a JavaFX TextArea, and simply open cmd.exe and put everything there. Therefore I deleted everything about the TextArea and resized the window

I also made some progress regarding the actual bruteforce. I finished the movable.sed calculations, and I included all of the required files for calling the C/C++ programs that do the bruteforce. Admittedly, I only included the GPU bruteforcer because I want to wait for Myria to finish her intel AVX bruteforcer before I include a CPU bruteforcer. I'm hoping that it will take in the required movable.sed and ID0 as arguments rather than a file so that I won't need to export a file (not that it is hard but i'd rather not do it)

I said "sort of working bfcl" because, while it seems to dump a correct movable.sed in the working directory I have set it to, it is sort of wonky with the "Press any key to continue" prompts. Needs fixing

other than that, progress is good!