Add ability to open movable_part1.sed files as well
Honestly I would prefer if everything used the txt format, but eh seems like no one has adopted it at all
It will remain in the code however!

By the way, if anyone wants to use this, IT WORKS JUST FINE! Go ahead and clone it and use it as you want!